requirements.txt
google-genai
pypdf
python-dotenv
tqdm
aiolimiter
tenacity
aiosqlite
